Move email helper functions to the new utils/domain.ts module
